refactor: search input as a standalone component (#2032) (#2054)
* refactor: search input as a standalone component

* revert: unnecessary update

* fix: build

* fix: test

* fix: placeholder

* fix: search page
