+++ packages/shared/src/components/feeds/FeedContainer.tsx
@@ -111,7 +111,6 @@ export const FeedContainer = ({
       <ScrollToTopButton />
       <div className="flex flex-col pt-2 laptopL:mx-auto w-full" style={style}>
         {!inlineHeader && header}
-
         <div
           className={classNames(
             'relative mx-auto w-full',
@@ -123,15 +122,7 @@ export const FeedContainer = ({
           data-testid="posts-feed"
         >
           {inlineHeader && header}
-
-          {showSearch && (
-            <SearchBar
-              className={{ container: 'mb-8' }}
-              inputId="search"
-              completedTime="12:12"
-            />
-          )}
-
+          {showSearch && <SearchBar className={{ container: 'mb-8' }} />}
           <div
             className={classNames(
               'grid',
